+++ changmin-todolist/server/routes/user.js
@@ -1,0 +1,5 @@
+const express = require("express");
+const router = express.Router();
+const userController = require("../controllers/userController");
+
+module.exports = router;
